Add lists/finnix-firmware.list.chroot
In past live-build, ALL packages with firmware were included, but
this was an unintended edge case.  With that fixed, we want to
explicitly define which firmware packages we want.  Here we install
most, including all known ethernet/wifi firmware, so if the user
needs a more esoteric firmware package, they can get it after boot.