Add `strings` reimplementation
I had avoided including `strings` because, while it's incredibly
useful, it's the only desirable utility in the binutils package,
which otherwise includes a bunch of compilation-related utilities,
and the package itself is quite large.

So in the words of the "my mechanics" YouTube channel, I make a
new one! It's not a 100% complete reimplementation of GNU
binutils' strings, but is fine for casual binary checking.

This is also set up so if you do `apt install binutils`, its
strings will take precedence over the Python version.